Tiles generated by adjacent corner Coordinates

+++ Dizignit.Core/MapCordinate.cs
@@ -2,12 +2,12 @@
 {
     public class MapCordinate
     {
-        public MapCordinate(decimal latitude, decimal longitude)
+        public MapCordinate(double latitude, double longitude)
         {
             Latitude = latitude;
             Longitude = longitude;
         }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
